# backend/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ITeacher": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"roles": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"type": {
						"type": "string"
					},
					"courses": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"email",
					"password",
					"roles",
					"type",
					"courses"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICourseQuestion": {
				"properties": {
					"description": {
						"type": "string"
					},
					"answers": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"correctAnswer": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"description",
					"answers",
					"correctAnswer"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICourse": {
				"properties": {
					"teacher": {
						"$ref": "#/components/schemas/ITeacher"
					},
					"name": {
						"type": "string"
					},
					"chapters": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"content": {
						"type": "string"
					},
					"questions": {
						"items": {
							"$ref": "#/components/schemas/ICourseQuestion"
						},
						"type": "array"
					}
				},
				"required": [
					"teacher",
					"name",
					"chapters",
					"content",
					"questions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CourseCreateModel": {
				"properties": {
					"name": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"chapters": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"questions": {
						"items": {
							"$ref": "#/components/schemas/ICourseQuestion"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"content",
					"chapters",
					"questions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CourseWithSubscribers": {
				"properties": {
					"teacher": {
						"$ref": "#/components/schemas/ITeacher"
					},
					"name": {
						"type": "string"
					},
					"chapters": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"content": {
						"type": "string"
					},
					"questions": {
						"items": {
							"$ref": "#/components/schemas/ICourseQuestion"
						},
						"type": "array"
					},
					"subscribers": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"teacher",
					"name",
					"chapters",
					"content",
					"questions",
					"subscribers"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Mario Deaconescu"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/courses/subscribedCourses": {
			"get": {
				"operationId": "GetSubscribedCourses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ICourse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Course"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/courses/find": {
			"get": {
				"operationId": "Find",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ICourse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Course"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "chapters",
						"required": true,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					}
				]
			}
		},
		"/courses/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ICourse"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Course"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CourseCreateModel"
							}
						}
					}
				}
			}
		},
		"/courses/myCourses": {
			"get": {
				"operationId": "MyCourses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CourseWithSubscribers"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Course"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}